mcp-server-1/pyproject.toml
[project]
name = "mcp-server-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.9.0",
    "psycopg2-binary>=2.9.10",
]
